+++ public/config.js
@@ -1,0 +1,10 @@
+/**
+ * config.js
+ * 
+ * Archivo de configuración centralizada para la aplicación.
+ * En producción, se recomienda que la clave secreta se cargue desde variables de entorno u otro mecanismo seguro.
+ */
+export const CONFIG = {
+    SECRET_KEY: "mySuperSecretKey123" // Cambia este valor en producción mediante variables de entorno o un archivo seguro.
+  };
+
