+++ public/index.html
@@ -16,7 +16,6 @@
       <nav>
         <ul>
           <li><a href="#" id="menu-gestion-periodos" aria-label="Gestión de periodos">Gestión de periodos</a></li>
-          <!-- Se pueden agregar más opciones en el futuro -->
         </ul>
       </nav>
     </aside>
@@ -63,6 +62,7 @@
   <footer>
     <p>&copy; 2025 Periodos App. Todos los derechos reservados.</p>
   </footer>
-  <script src="script.js"></script>
+  <!-- Se actualiza la llamada al archivo index.js como módulo -->
+  <script type="module" src="index.js"></script>
 </body>
 </html>
