+++ public/styles.css
@@ -1,180 +1,179 @@
 /* Global Styles */
 body {
-    font-family: 'Roboto', sans-serif;
-    background-color: #f4f4f4;
-    margin: 0;
-    padding: 0;
-    color: #333;
-  }
-  
-  /* Contenedor principal que engloba el sidebar y el contenido */
-  #app-container {
-    display: flex;
-    min-height: 100vh;
-  }
-  
-  /* Menú lateral */
+  font-family: 'Roboto', sans-serif;
+  background-color: #f4f4f4;
+  margin: 0;
+  padding: 0;
+  color: #333;
+}
+
+/* Contenedor principal que engloba el sidebar y el contenido */
+#app-container {
+  display: flex;
+  min-height: 100vh;
+}
+
+/* Menú lateral */
+#sidebar {
+  background-color: #2c3e50;
+  color: #ecf0f1;
+  width: 220px;
+  padding: 20px;
+  box-sizing: border-box;
+}
+
+#sidebar nav ul {
+  list-style: none;
+  padding: 0;
+}
+
+#sidebar nav ul li {
+  margin-bottom: 15px;
+}
+
+#sidebar nav ul li a {
+  color: #ecf0f1;
+  text-decoration: none;
+  font-weight: bold;
+}
+
+/* Contenido principal */
+#main-content {
+  flex: 1;
+  padding: 20px;
+  box-sizing: border-box;
+}
+
+header {
+  background-color: #007BFF;
+  color: #fff;
+  padding: 20px;
+  text-align: center;
+}
+
+header h1 {
+  margin: 0;
+  font-size: 2rem;
+}
+
+header p {
+  margin: 10px 0 0;
+  font-size: 1rem;
+}
+
+main {
+  max-width: 900px;
+  margin: 20px auto;
+  padding: 0 15px;
+}
+
+/* Secciones de Vistas */
+section {
+  background: #fff;
+  padding: 20px;
+  margin-bottom: 20px;
+  border-radius: 8px;
+  box-shadow: 0 2px 5px rgba(0,0,0,0.1);
+}
+
+.hidden {
+  display: none;
+}
+
+h2 {
+  margin-top: 0;
+  font-size: 1.5rem;
+  border-bottom: 2px solid #007BFF;
+  padding-bottom: 5px;
+}
+
+/* Card Container for Listados */
+.card-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 15px;
+  margin-bottom: 15px;
+}
+
+/* Tarjetas para Periodos y Meses */
+.card {
+  background: #f9f9f9;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  padding: 15px;
+  flex: 1 1 250px;
+  box-shadow: 0 1px 3px rgba(0,0,0,0.1);
+  transition: transform 0.3s ease, opacity 0.3s ease;
+}
+
+.card:hover {
+  transform: translateY(-5px);
+  opacity: 0.95;
+}
+
+/* Ejemplo de estilo para destacar Ganancia Real */
+.ganancia-real {
+  font-weight: bold;
+  color: #28a745;
+}
+
+/* Botones */
+button {
+  background-color: #007BFF;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  padding: 10px 15px;
+  margin: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+button:hover {
+  background-color: #0056b3;
+}
+
+/* Acciones en Vista de Meses y Navegación */
+.acciones-meses,
+.navegacion-detalle,
+.navegacion-resultados {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+  margin-top: 10px;
+}
+
+/* Inputs en tarjetas (para meses) */
+.card input[type="text"],
+.card input[type="number"] {
+  width: 100%;
+  padding: 8px;
+  margin-top: 5px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+
+/* Results Section */
+#metrics p {
+  margin: 5px 0;
+}
+
+/* Footer */
+footer {
+  background-color: #333;
+  color: #fff;
+  text-align: center;
+  padding: 15px;
+  font-size: 0.9rem;
+}
+
+/* Responsive adjustments */
+@media (max-width: 600px) {
   #sidebar {
-    background-color: #2c3e50;
-    color: #ecf0f1;
-    width: 220px;
-    padding: 20px;
-    box-sizing: border-box;
+    width: 180px;
+    padding: 15px;
   }
-  
-  #sidebar nav ul {
-    list-style: none;
-    padding: 0;
-  }
-  
-  #sidebar nav ul li {
-    margin-bottom: 15px;
-  }
-  
-  #sidebar nav ul li a {
-    color: #ecf0f1;
-    text-decoration: none;
-    font-weight: bold;
-  }
-  
-  /* Contenido principal */
-  #main-content {
-    flex: 1;
-    padding: 20px;
-    box-sizing: border-box;
-  }
-  
-  header {
-    background-color: #007BFF;
-    color: #fff;
-    padding: 20px;
-    text-align: center;
-  }
-  
-  header h1 {
-    margin: 0;
-    font-size: 2rem;
-  }
-  
-  header p {
-    margin: 10px 0 0;
-    font-size: 1rem;
-  }
-  
-  main {
-    max-width: 900px;
-    margin: 20px auto;
-    padding: 0 15px;
-  }
-  
-  /* Secciones de Vistas */
-  section {
-    background: #fff;
-    padding: 20px;
-    margin-bottom: 20px;
-    border-radius: 8px;
-    box-shadow: 0 2px 5px rgba(0,0,0,0.1);
-  }
-  
-  .hidden {
-    display: none;
-  }
-  
-  h2 {
-    margin-top: 0;
-    font-size: 1.5rem;
-    border-bottom: 2px solid #007BFF;
-    padding-bottom: 5px;
-  }
-  
-  /* Card Container for Listados */
-  .card-container {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 15px;
-    margin-bottom: 15px;
-  }
-  
-  /* Tarjetas para Periodos y Meses */
   .card {
-    background: #f9f9f9;
-    border: 1px solid #ccc;
-    border-radius: 8px;
-    padding: 15px;
-    flex: 1 1 250px;
-    box-shadow: 0 1px 3px rgba(0,0,0,0.1);
-    transition: transform 0.3s ease, opacity 0.3s ease;
+    flex: 1 1 100%;
   }
-  
-  .card:hover {
-    transform: translateY(-5px);
-    opacity: 0.95;
-  }
-  
-  /* Ejemplo de estilo para destacar Ganancia Real */
-  .ganancia-real {
-    font-weight: bold;
-    color: #28a745;
-  }
-  
-  /* Botones */
-  button {
-    background-color: #007BFF;
-    color: #fff;
-    border: none;
-    border-radius: 4px;
-    padding: 10px 15px;
-    margin: 5px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-  }
-  
-  button:hover {
-    background-color: #0056b3;
-  }
-  
-  /* Acciones en Vista de Meses y Navegación */
-  .acciones-meses,
-  .navegacion-detalle,
-  .navegacion-resultados {
-    display: flex;
-    gap: 10px;
-    flex-wrap: wrap;
-    margin-top: 10px;
-  }
-  
-  /* Inputs en tarjetas (para meses) */
-  .card input[type="text"],
-  .card input[type="number"] {
-    width: 100%;
-    padding: 8px;
-    margin-top: 5px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  /* Results Section */
-  #metrics p {
-    margin: 5px 0;
-  }
-  
-  /* Footer */
-  footer {
-    background-color: #333;
-    color: #fff;
-    text-align: center;
-    padding: 15px;
-    font-size: 0.9rem;
-  }
-  
-  /* Responsive adjustments */
-  @media (max-width: 600px) {
-    #sidebar {
-      width: 180px;
-      padding: 15px;
-    }
-    .card {
-      flex: 1 1 100%;
-    }
-  }
-  
+}
